Add create pack

+++ Fayora.Domain/Entities/GuideModule/TourGuide.cs
@@ -32,7 +32,7 @@ public class TourGuide : GuideAccountBase
         PricingUnit = pricingUnit;
         BaseRate = baseRate;
         YearsOfExperience = yearsOfExperience;
-        AverageRating = 0f;
+        AverageRating = 0m;
         Status = GuideStatus.Pending;
         IsAvailableForBooking = false;
         CompletedToursCount = 0;
@@ -70,7 +70,7 @@ public class TourGuide : GuideAccountBase
     }
 
 
-    public void UpdateRating(float newRating)
+    public void UpdateRating(decimal newRating)
     {
         AverageRating = (AverageRating * CompletedToursCount + newRating) / (CompletedToursCount + 1);
 
@@ -125,7 +125,7 @@ public class TourGuide : GuideAccountBase
             _cityIds.Remove(cityId);
     }
 
-    public void AddReview(float newRating)
+    public void AddReview(decimal newRating)
     {
         AverageRating = ((AverageRating * ReviewCount) + newRating) / (ReviewCount + 1);
         ReviewCount++;
